charttomd/pyproject.toml
[project]
name = "charttomd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.11",
    "azure-identity>=1.19.0",
    "pydantic-settings>=2.7.0",
    "pydantic>=2.10.4",
    "tenacity>=9.0.0",
    "typer>=0.15.1",
    "mypy>=1.14.0",
]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
